refactor: split scoped tools by resource and lead with canonical tools
